Add save and loading pokemon party states
To test this, start a battle with an opponent, then at the start of the fight (when you can see his pokemon) press Q to save his pokemon party. After defeating him, go to another opponent and before talking to him, press W. After the battle starts, you'll see that the trainer's lineup is the same as the opponent you pressed Q for. Press E before starting your first move and fight the trainer as normally.